feat(all): new router configuration strategy
BREAKING CHANGE. Any view model that wants  a router for configuration
(ether app level or child) must implement a configureRouter callback,
which will receive the config object and the router instance. There is
now no requirement for a “router” property and it is much clearer when
a view model wants to add a new level to the router hierarchy
(configureRouter) vs. when it wants to use an existing router for
navigation purposes.