feat(router): observables in lifecycle
Given that Observables are moving through committee to become a web
standard, it would be of use to support that standard similarly to how
Promises are supported.

Like with promises, if an observable (really, anything with a subscribe
function) is returned from the activate lifecycle event, the router will
wait until there's data present before finishing the routing process.