fix(router): navigation model explicit href normalization
This allows explicit hrefs to be set on the nav configuration without
needing to consider push state or not. The nav model will account for
that automatically. This works in conjunction with a fix in the route
recognizer which now takes into account explicit href when generating
routes. Since generation also normalizes, this allows for consistent
configuration.