doc(typescript): fix mapUnknownRoutes Typescript compile error
The TypeScript examples that shows how to setup a function to map unknown routes has an error in it. At the moment, it returns an object containing a `moduleId` string. However, this leads to a compile error, due to the `RouteConfig` field also requiring the `route` field to be specified.