Test: Remove obsolete activation test
I removed a test that would erroneously fail. The test was checking that a plan with a childNavigationInstruction property that returned true or false on canDeactivate would continue or cancel respectively. The mock childNavigationInstruction was missing the required childRouter property, and so it was giving erroneous results.

The following test, 'with router and currentInstruction', actually includes coverage for the very same feature. Therefore, removing the erroneous test was sufficient.