Fix: Reorder pipeline steps
In the Pipeline, LoadRouteStep was being called after CanDeactivatePreviousStep. There is a bug in the CanDeactivatePreviousStep and DeactivatePreviousStep logic that is only apparent after the LoadRouteStep is run. Therefore, the bug was only surfacing in DeactivatePreviousStep. However, I believe that LoadRouteStep makes more sense being called before CanDeactivatePreviousStep, and in fact making this change makes the bug more apparent.