Fix: Call addPreviousDeactivatable when deactivating a parent with an activated child
Before, if a viewPortPlan had a childNavigationInstruction (which is added to the plan in the LoadRouteStep), findDeactivatable would call itself with childNavigationInstruction.plan. However, in the case where the plan was built off of a parentInstruction, the childNavigationInstruction.plan would be lacking a prevComponent property, rightly so, I believe.

Instead, calling addPreviousDeactivatable reaches into the childRouter on the component, if present, gets all currently activated view models, and queues all of the can/deactivate callbacks.

I couldn't find the use case for calling findDeactivatable recursively. I removed it and all working tests passed.