feat(router): Enhance optional view port feature
I've added a little bit more robust handling on the router-configuration and router objects for configuring the default view port configuration, and I've also added typedocs for this. I've rewritten the navigation plan to be a little more straightforward and consistent with the previous implementation. I removed the configureExplicitViewports function because I'm concerned it will bring low value and high confusion.