build(pipeline-provider): Improve PipelineProvider types
The use of Slots and Steps in the Pipeline/Provider is a bit sloppy and not clear. Both cases, Slots in the PipelineProvider and Steps in the Pipeline, support either slots or steps respectively or constructors for the associated steps. Adding these types allows us to remove typecasting workarounds. There is one place where some duck typing was used and typecasting is required. I've added a comment describing why.