chore(interfaces): Tighten up the RouteConfigSpecifier typing
Expected a RouteConfig before, but that's not true. It wasn't interested in the path, the only required part of the RouteConfig type, just the result, whether its a redirect, viewModel, or moduleId. Created a new and more specific type for this purpose so we don't mangle up the RouteConfig and obscure its purpose.